fix(broker): use single_process for sanic
In 22.9 sanic underwent a large overhaul to change the way it starts up
worker processes, such that it will always *spawn* a process even when
workers=1. This seems to cause significant issues with the way comet
functions. A full fix would involve cleaning up the logic and state to ensure a
clean spawning, however, as we've never supported more than one worker,
it's much easier just to set sanic in single_process mode.